+++ CHANGELOG.md
@@ -5,3 +5,13 @@
 *  Update theme to default to white for header background
 *  Add `md-primary` styling for header background and selected events. You can add the `md-primary` class to the `me-event-calendar` directive.
 *  Fix event sorting in month view to fit more events in a day
+
+
+<a name="0.0.6"></a>
+## 0.0.6 (2016-09-07)
+
+*  Add Today button directive
+*  Add `md-create-event-click` attribute that will trigger on month cell click. This passes a `$date` paramater
+*  Add `md-show-create-link` attribute to display a create link on cell hover
+*  Add `md-create-disabled` to disable and hide create
+*  Fix event display on last day of week